Add daemon mode and command line arguments.
Added command line arguments and a daemon mode for the script.

Added idleloop method which sits and waits for either the playlist to
change (then update the playlist) or a song to change at which it calls
another new method that returns True if the current song is the last
song in the album, if True then it will wait for it to change then play
a random album.