Update shuffle method to no longer assign new ids to songs.
The old shuffle method would completely clear the playlist then add
albums in the new shuffled order, this in turn assiged each song a new
id. This made so we can no longer identify a song precisely after the
playlist has been shuffled.

The insert_album() method has been updated to actually insert an album
this, by taking an album and an optional pos argument and moves the
album to that position.

With this change we no longer need to keep track of the last song since
mpd will remember, and will not change its state when shuffling. This
makes shuffling albums completely seamless.